sdk stage 1: split gesture vocabulary into TapQGestureContracts
Move the gesture/input vocabulary (HeadGesture, VoiceCommand, TapCommand,
TiltCommand, swipe commands, provider protocols) and the diagnostics
contract out of TapQContracts into a new dependency-free portable target.
TapQContracts re-exports it so every existing import site compiles
unchanged; TapQDetectionBaseline now depends only on TapQGestureContracts,
making detection embeddable without the agent/approval domain.
