+++ Sources/TapQGestureContracts/Diagnostics.swift
